refactor: extract CLI command handlers from main.rs into separate modules
Split the 4098-line main.rs into focused modules under commands/:
- commands/apply.rs: apply, apply_from_plan, execute_effects, finalize_apply, drift detection
- commands/destroy.rs: destroy command with timeout/wait logic
- commands/plan.rs: plan command, PlanFile/CurrentStateEntry types
- commands/state.rs: state refresh, bucket delete, force-unlock, map_lock_error
- commands/validate.rs: validate command
- commands/fmt.rs: format command with diff display
- commands/lint.rs: lint command
- commands/module.rs: module info command
- commands/mod.rs: shared validate_and_resolve pipeline

main.rs now contains only CLI argument parsing, command dispatch, and tests.

Part of #777
