+++ carina-cli/src/commands/lint.rs
@@ -1,0 +1,110 @@
+use std::collections::{HashMap, HashSet};
+use std::fs;
+use std::path::PathBuf;
+
+use colored::Colorize;
+
+use carina_core::config_loader::{find_crn_files_in_dir, get_base_dir, load_configuration};
+use carina_core::lint::{find_list_literal_attrs, list_struct_attr_names};
+use carina_core::module_resolver;
+use carina_core::provider::{self as provider_mod};
+
+use crate::error::AppError;
+use crate::wiring::{get_schemas, provider_factories};
+
+/// A lint warning with file, line, and message info.
+struct LintWarning {
+    file: PathBuf,
+    line: usize,
+    message: String,
+}
+
+pub fn run_lint(path: &PathBuf) -> Result<(), AppError> {
+    let mut parsed = load_configuration(path)?.parsed;
+
+    let base_dir = get_base_dir(path);
+
+    // Resolve modules
+    module_resolver::resolve_modules(&mut parsed, base_dir)
+        .map_err(|e| format!("Module resolution error: {}", e))?;
+
+    let factories = provider_factories();
+    let schemas = get_schemas();
+
+    // Collect source texts for each .crn file
+    let source_texts: Vec<(PathBuf, String)> = if path.is_file() {
+        let content = fs::read_to_string(path)
+            .map_err(|e| format!("Failed to read {}: {}", path.display(), e))?;
+        vec![(path.clone(), content)]
+    } else if path.is_dir() {
+        let files = find_crn_files_in_dir(path)?;
+        let mut texts = Vec::new();
+        for file in files {
+            let content = fs::read_to_string(&file)
+                .map_err(|e| format!("Failed to read {}: {}", file.display(), e))?;
+            texts.push((file, content));
+        }
+        texts
+    } else {
+        return Err(AppError::Config(format!(
+            "Path not found: {}",
+            path.display()
+        )));
+    };
+
+    // Collect all List<Struct> attribute names from schemas of parsed resources
+    // and build a map of attr_name -> block_name for lint suggestions
+    let mut all_list_struct_attrs: HashSet<String> = HashSet::new();
+    let mut block_name_suggestions: HashMap<String, String> = HashMap::new();
+    for resource in &parsed.resources {
+        let schema_key = provider_mod::schema_key_for_resource(&factories, resource);
+        if let Some(schema) = schemas.get(&schema_key) {
+            all_list_struct_attrs.extend(list_struct_attr_names(schema));
+            for (attr_name, attr_schema) in &schema.attributes {
+                if let Some(bn) = &attr_schema.block_name {
+                    block_name_suggestions.insert(attr_name.clone(), bn.clone());
+                }
+            }
+        }
+    }
+
+    // Scan each source file for list literal usage of List<Struct> attributes
+    let mut warnings: Vec<LintWarning> = Vec::new();
+
+    for (file_path, source) in &source_texts {
+        let hits = find_list_literal_attrs(source, &all_list_struct_attrs);
+        for (attr_name, line) in hits {
+            let suggested_name = block_name_suggestions
+                .get(&attr_name)
+                .map(|s| s.as_str())
+                .unwrap_or(&attr_name);
+            warnings.push(LintWarning {
+                file: file_path.clone(),
+                line,
+                message: format!(
+                    "Prefer block syntax for '{}'. Use `{} {{ ... }}` instead of `{} = [{{ ... }}]`.",
+                    attr_name, suggested_name, attr_name
+                ),
+            });
+        }
+    }
+
+    if warnings.is_empty() {
+        println!("{}", "No lint warnings found.".green().bold());
+        Ok(())
+    } else {
+        for w in &warnings {
+            eprintln!(
+                "{} {}:{}  {}",
+                "warning:".yellow().bold(),
+                w.file.display(),
+                w.line,
+                w.message
+            );
+        }
+        Err(AppError::Validation(format!(
+            "Found {} lint warning(s).",
+            warnings.len()
+        )))
+    }
+}
